refactor: extract paragraph and recenter to src/editor-ops/
Move moveToNextParagraph (with Direction enum) and recenterToBottom
out of the plugin class into pure free functions under src/editor-ops/.
Both were already effectively pure; this just removes the 'this.' bind.

buildCommands now calls these free functions directly instead of going
through plugin-class methods.

Part of Task 1.8 (extract editor-ops modules).
